+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - 🚨 Update to Assembly v1.
 
+## 4.2.3
+
+- Fix casing for `DownloadButton` with GeoJSON filetype. [#484](https://github.com/mapbox/dr-ui/pull/484)
+
 ## 4.2.2
 
 - Fix issue in `NavigationAccordion` to handle pages that do not have a parent. [#479](https://github.com/mapbox/dr-ui/pull/479)